pyproject.toml
[project]
name = "fastapidemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich[toml]>=0.9.1",
    "fastapi[standard]>=0.116.1",
    "passlib>=1.7.4",
    "bcrypt==4.0.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "tortoise-orm[aiomysql]>=0.25.1",
    "cryptography>=45.0.6",
    "redis>=6.4.0",
]

[[tool.uv.index]]
url="https://mirrors.aliyun.com/pypi/simple/"
default=true

[tool.aerich]
tortoise_orm = "src.databases.mysql_session.db_config"
location = "./migrations"
src_folder = "./."
